REGLERA MERA

+++ src/system/mimer.c
@@ -40,7 +40,7 @@ int8_t parallelR = 0x0c;
 /*reflex-sensor*/
 uint8_t reflex_previous = 0x00;
 uint8_t reflex_current = 0x00;
-uint8_t segments_turned = 0x00;
+uint16_t segments_turned = 0x00;
 uint8_t MR_Reflex = 0x00;
 
 /*GYRO*/
